Bugfix: coordinates should be floats, shouldn't they?

+++ daft_scraper/listing.py
@@ -58,7 +58,7 @@ class ListingPoint(Schema):
         unknown = INCLUDE
 
     point_type = fields.Str(data_key="type")
-    coordinates = fields.List(fields.Int())
+    coordinates = fields.List(fields.Float())
 
 
 class ListingPRS(Schema):
